+++ packages/vsce/__tests__/__unit__/extending/CICSExtenderApiConfig.unit.test.ts
@@ -20,6 +20,7 @@ describe("CICS Extender Api Tests", () => {
 
   it("should return supported resources configuration", () => {
     const config: IExtensionAPI = CICSExtenderApiConfig.getAPI();
-    expect(config).toHaveProperty('resources.supportedResources', []);
+    expect(config).toHaveProperty('resources.supportedResources');
+    expect(config.resources.supportedResources).toHaveLength(9);
   });
 });