+++ packages/vsce-api/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the IBM CICS for Zowe Explorer API will be documented in this file.
 
+## Recent Changes
+
+- Populate supported resources list
+
 ## `6.9.4`
 
 - Return list of supported resources - currently empty